+++ src/utils/H2DE_color_hsv.cpp
@@ -14,22 +14,27 @@ H2DE_ColorHSV::operator H2DE_ColorRGB() const {
         r_f = c;
         g_f = x;
         b_f = 0;
+
     } else if (h >= 60 && h < 120) {
         r_f = x;
         g_f = c;
         b_f = 0;
+
     } else if (h >= 120 && h < 180) {
         r_f = 0;
         g_f = c;
         b_f = x;
+
     } else if (h >= 180 && h < 240) {
         r_f = 0;
         g_f = x;
         b_f = c;
+
     } else if (h >= 240 && h < 300) {
         r_f = x;
         g_f = 0;
         b_f = c;
+
     } else {
         r_f = c;
         g_f = 0;
@@ -58,8 +63,14 @@ H2DE_ColorHSV H2DE_ColorHSV::addHue(float hue) const {
     H2DE_ColorHSV hsv = *this;
     hsv.h += hue;
 
-    while (hsv.h >= 360) hsv.h -= 360.0f;
-    while (hsv.h < 0.0f) hsv.h += 360.0f;
+    while (hsv.h >= 360) {
+        hsv.h -= 360.0f;
+    }
+
+    while (hsv.h < 0.0f) {
+        hsv.h += 360.0f;
+    }
+
     return hsv;
 }
 
@@ -67,8 +78,13 @@ H2DE_ColorHSV H2DE_ColorHSV::addSaturation(float saturation) const {
     H2DE_ColorHSV hsv = *this;
     hsv.s += saturation;
 
-    if (hsv.s > 1.0f) hsv.s = 1.0f;
-    else if (hsv.s < 0.0f) hsv.s = 0.0f;
+    if (hsv.s > 1.0f) {
+        hsv.s = 1.0f;
+
+    } else if (hsv.s < 0.0f) {
+        hsv.s = 0.0f;
+    }
+
     return hsv;
 }
 
@@ -76,8 +92,13 @@ H2DE_ColorHSV H2DE_ColorHSV::addValue(float value) const {
     H2DE_ColorHSV hsv = *this;
     hsv.v += value;
 
-    if (hsv.v > 1.0f) hsv.v = 1.0f;
-    else if (hsv.v < 0.0f) hsv.v = 0.0f;
+    if (hsv.v > 1.0f) {
+        hsv.v = 1.0f;
+    
+    } else if (hsv.v < 0.0f) {
+        hsv.v = 0.0f;
+    }
+
     return hsv;
 }
 
@@ -99,8 +120,14 @@ H2DE_ColorHSV H2DE_ColorHSV::multiplyHue(float multiplier) const {
     H2DE_ColorHSV hsv = *this;
     hsv.h *= multiplier;
 
-    while (hsv.h >= 360) hsv.h -= 360.0f;
-    while (hsv.h < 0.0f) hsv.h += 360.0f;
+    while (hsv.h >= 360) {
+        hsv.h -= 360.0f;
+    }
+
+    while (hsv.h < 0.0f) {
+        hsv.h += 360.0f;
+    }
+
     return hsv;
 }
 
@@ -108,8 +135,13 @@ H2DE_ColorHSV H2DE_ColorHSV::multiplySaturation(float multiplier) const {
     H2DE_ColorHSV hsv = *this;
     hsv.s *= multiplier;
 
-    if (hsv.s > 1.0f) hsv.s = 1.0f;
-    else if (hsv.s < 0.0f) hsv.s = 0.0f;
+    if (hsv.s > 1.0f) {
+        hsv.s = 1.0f;
+
+    } else if (hsv.s < 0.0f) {
+        hsv.s = 0.0f;
+    }
+
     return hsv;
 }
 
@@ -117,8 +149,13 @@ H2DE_ColorHSV H2DE_ColorHSV::multiplyValue(float multiplier) const {
     H2DE_ColorHSV hsv = *this;
     hsv.v *= multiplier;
 
-    if (hsv.v > 1.0f) hsv.v = 1.0f;
-    else if (hsv.v < 0.0f) hsv.v = 0.0f;
+    if (hsv.v > 1.0f) {
+        hsv.v = 1.0f;
+
+    } else if (hsv.v < 0.0f) {
+        hsv.v = 0.0f;
+    }
+    
     return hsv;
 }
 
@@ -133,4 +170,9 @@ H2DE_ColorHSV H2DE_ColorHSV::divideSaturation(float divider) const {
 
 H2DE_ColorHSV H2DE_ColorHSV::divideValue(float divider) const {
     return multiplyValue(1 / divider);
+}
+
+// METHODS
+const bool H2DE_ColorHSV::isVisible() const {
+    return a != 0;
 }